chatScreen [nfc]: Inline single-use `Chat` into `ChatScreen`.
