+++ src/chat/ChatScreen.js
@@ -4,23 +4,45 @@ import { View, StyleSheet } from 'react-native';
 import type { NavigationScreenProp } from 'react-navigation';
 import { ActionSheetProvider } from '@expo/react-native-action-sheet';
 
-import type { Context, Narrow } from '../types';
-import { OfflineNotice, ZulipStatusBar } from '../common';
-import Chat from './Chat';
+import { connect } from '../react-redux';
+import type { Context, Dispatch, Fetching, Narrow } from '../types';
+import { KeyboardAvoider, OfflineNotice, ZulipStatusBar } from '../common';
 import ChatNavBar from '../nav/ChatNavBar';
+
+import MessageList from '../webview/MessageList';
+import NoMessages from '../message/NoMessages';
+import ComposeBox from '../compose/ComposeBox';
+import UnreadNotice from './UnreadNotice';
+
+import styles from '../styles';
+
+import { canSendToNarrow } from '../utils/narrow';
+import { getLoading } from '../directSelectors';
+import { getFetchingForNarrow } from './fetchingSelectors';
+import { getShownMessagesForNarrow } from './narrowsSelectors';
+
+type SelectorProps = {|
+  fetching: Fetching,
+  haveNoMessages: boolean,
+  loading: boolean,
+|};
 
 type Props = $ReadOnly<{|
   navigation: NavigationScreenProp<{ params: {| narrow: Narrow |} }>,
+  dispatch: Dispatch,
+
+  ...SelectorProps,
 |}>;
 
-const styles = StyleSheet.create({
-  /** A workaround for #3089, by letting us put Chat (with MessageList) first. */
+const componentStyles = StyleSheet.create({
+  /** A workaround for #3089, by letting us put MessageList first. */
   reverse: {
+    flex: 1,
     flexDirection: 'column-reverse',
   },
 });
 
-export default class ChatScreen extends PureComponent<Props> {
+class ChatScreen extends PureComponent<Props> {
   context: Context;
 
   static contextTypes = {
@@ -29,12 +51,27 @@ export default class ChatScreen extends PureComponent<Props> {
 
   render() {
     const { styles: contextStyles } = this.context;
-    const { narrow } = this.props.navigation.state.params;
+    const { fetching, haveNoMessages, loading, navigation } = this.props;
+    const { narrow } = navigation.state.params;
+
+    const isFetching = fetching.older || fetching.newer || loading;
+    const showMessagePlaceholders = haveNoMessages && isFetching;
+    const sayNoMessages = haveNoMessages && !isFetching;
+    const showComposeBox = canSendToNarrow(narrow) && !showMessagePlaceholders;
 
     return (
       <ActionSheetProvider>
-        <View style={[contextStyles.screen, styles.reverse]}>
-          <Chat narrow={narrow} />
+        <View style={[contextStyles.screen, componentStyles.reverse]}>
+          <KeyboardAvoider style={styles.flexed} behavior="padding">
+            <View style={styles.flexed}>
+              <View style={componentStyles.reverse}>
+                <MessageList narrow={narrow} showMessagePlaceholders={showMessagePlaceholders} />
+                {sayNoMessages && <NoMessages narrow={narrow} />}
+                <UnreadNotice narrow={narrow} />
+              </View>
+              {showComposeBox && <ComposeBox narrow={narrow} />}
+            </View>
+          </KeyboardAvoider>
           <OfflineNotice />
           <ChatNavBar narrow={narrow} />
           <ZulipStatusBar narrow={narrow} />
@@ -43,3 +80,10 @@ export default class ChatScreen extends PureComponent<Props> {
     );
   }
 }
+
+export default connect<SelectorProps, _, _>((state, props) => ({
+  loading: getLoading(state),
+  fetching: getFetchingForNarrow(state, props.navigation.state.params.narrow),
+  haveNoMessages:
+    getShownMessagesForNarrow(state, props.navigation.state.params.narrow).length === 0,
+}))(ChatScreen);
